v2.3 Migration to Staging

+++ src/entities/supplemental-operating.entity.ts
@@ -21,7 +21,7 @@ export class SupplementalOperating extends BaseEntity {
 
   @Column({
     name: 'rpt_period_id',
-    type: 'numeric',
+    type: 'bigint',
     transformer: new NumericColumnTransformer(),
   })
   reportPeriodId: number;
